refactor(frontend): Move main components to new folder
