+++ frontend/src/app/components/about/about.component.scss
